fix merge

+++ engine/wii/control.c
@@ -73,7 +73,7 @@ void reset()
 	hwbutton = WII_RESET;
 }
 
-void wiimote_poweroff(long int playernum)
+void wiimote_poweroff(int playernum)
 {
 	hwbutton = WII_SHUTDOWN;
 }
